pyproject.toml
[project]
name = "testbedimage"
authors = [
    {name = "Wolfgang Hotwagner", email = "wolfgang.hotwagner@ait.ac.at"},
]
description = "Fetch images from openstack and deploy it via SFTP."
readme = "README.md"
requires-python = ">=3.7"
keywords = ["Testbed", "AECID", "OpenStack", "SFTP", "Image"]
license = {text = "GPL-3.0"}
dependencies = [
    "rich",
    "pydantic ~= 1.10",
    "openstacksdk",
    "paramiko",
    "python-dateutil",
    "httpx"
]
dynamic = ["version"]

[project.scripts]
testbedimage = "testbedimage.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]
